requirements.txt
crispy-bootstrap4==2022.1
Django==4.2.5
django-allauth==0.57.0
mysqlclient==2.1.0
djangorestframework==3.14.0
django-rest-auth==0.9.5
django-cors-headers==4.3.1 
# djangorestframework-simplejwt==5.3.1 
# djoser==2.2.2 